+++ lib/call_sheet/transaction.rb
@@ -18,6 +18,20 @@ module CallSheet
     end
     alias_method :[], :call
 
+    def subscribe(listeners)
+      if listeners.is_a?(Hash)
+        listeners.each do |step_name, listener|
+          steps.detect { |step| step.step_name == step_name }.subscribe(listener)
+        end
+      else
+        steps.each do |step|
+          step.subscribe(listeners)
+        end
+      end
+
+      self
+    end
+
     private
 
     def assert_valid_options(options)